refactor: Rename emailMarketingStoreTypes to ...actionTypes
