+++ src/features/EmailMarketing/shared/InternalTools/components/ExamplePrompts/components/ExamplePromptWithNavigation/ExamplePrompt.test.tsx
@@ -1,7 +1,7 @@
 import {render, screen} from '@testing-library/react';
 import ExamplePrompt from './ExamplePrompt';
 import {mockDispatch} from '../../../../../../../../utils/mocks/mocks';
-import {EmailMarketingActionType} from '../../../../../../store/emailMarketingStoreTypes';
+import {EmailMarketingActionType} from '../../../../../../store/emailMarketingActionTypes';
 import {renderComponentWithState} from '../../../../../../store/emailMarketingStoreUtils';
 import {emailMarketingInitialState} from '../../../../../../store/emailMarketingInitialState';
 import userEvent from '@testing-library/user-event';
