+++ src/features/EmailMarketing/shared/TargetingBuilderHeader/components/GenerateWithAIForm/GenerateWithAIForm.test.tsx
@@ -1,7 +1,7 @@
 import {fireEvent, render, screen, waitFor} from '@testing-library/react';
 import GenerateWithAIForm from './GenerateWithAIForm';
 import {mockDispatch} from '../../../../../../utils/mocks/mocks';
-import {EmailMarketingActionType} from '../../../../store/emailMarketingStoreTypes';
+import {EmailMarketingActionType} from '../../../../store/emailMarketingActionTypes';
 import {renderComponentWithState} from '../../../../store/emailMarketingStoreUtils';
 import {emailMarketingInitialState} from '../../../../store/emailMarketingInitialState';
 import userEvent from '@testing-library/user-event';
