+++ src/features/EmailMarketing/shared/TriggerSelection/TriggerSelection.tsx
@@ -1,6 +1,6 @@
 import Button from '../../../shared/ui/components/Button/Button';
 import Sticker from '../../../shared/ui/components/Sticker/Sticker';
-import {EmailMarketingActionType} from '../../store/emailMarketingStoreTypes';
+import {EmailMarketingActionType} from '../../store/emailMarketingActionTypes';
 import {useEmailMarketingState} from '../../store/useEmailMarketingState';
 import {TriggerType} from '../emailMarketingTypes';
 import {triggers} from './triggers';
